Continue refactoring partitioned data reader and markers
