+++ tests/op_coreutils/partitioned/test_breakout.py
@@ -1,12 +1,10 @@
 import polars as pl
 
-
 from op_analytics.coreutils.partitioned.breakout import breakout_partitions
-from op_analytics.coreutils.partitioned.output import (
-    PartitionData,
-    OutputPartMeta,
+from op_analytics.coreutils.partitioned.partition import (
     PartitionColumn,
     PartitionColumns,
+    PartitionData,
 )
 
 
@@ -39,62 +37,50 @@ def test_breakout_partitions():
     for _ in outputs:
         actual.append(
             dict(
-                meta=_.meta,
-                full_path=_.meta.partitions.path,
+                partitions=_.partitions,
+                full_path=_.partitions.path,
                 nu_rows=len(_.df),
             )
         )
 
-    actual.sort(key=lambda x: x["meta"].partitions.path)
+    actual.sort(key=lambda x: x["full_path"])
     assert actual == [
         {
-            "meta": OutputPartMeta(
-                partitions=PartitionColumns(
-                    [
-                        PartitionColumn(name="dt", value="2024-01-01"),
-                        PartitionColumn(name="chain", value="base"),
-                    ]
-                ),
-                row_count=2,
+            "partitions": PartitionColumns(
+                cols=[
+                    PartitionColumn(name="dt", value="2024-01-01"),
+                    PartitionColumn(name="chain", value="base"),
+                ]
             ),
             "full_path": "dt=2024-01-01/chain=base",
             "nu_rows": 2,
         },
         {
-            "meta": OutputPartMeta(
-                partitions=PartitionColumns(
-                    [
-                        PartitionColumn(name="dt", value="2024-01-01"),
-                        PartitionColumn(name="chain", value="op"),
-                    ]
-                ),
-                row_count=2,
+            "partitions": PartitionColumns(
+                cols=[
+                    PartitionColumn(name="dt", value="2024-01-01"),
+                    PartitionColumn(name="chain", value="op"),
+                ]
             ),
             "full_path": "dt=2024-01-01/chain=op",
             "nu_rows": 2,
         },
         {
-            "meta": OutputPartMeta(
-                partitions=PartitionColumns(
-                    [
-                        PartitionColumn(name="dt", value="2024-01-02"),
-                        PartitionColumn(name="chain", value="base"),
-                    ]
-                ),
-                row_count=2,
+            "partitions": PartitionColumns(
+                cols=[
+                    PartitionColumn(name="dt", value="2024-01-02"),
+                    PartitionColumn(name="chain", value="base"),
+                ]
             ),
             "full_path": "dt=2024-01-02/chain=base",
             "nu_rows": 2,
         },
         {
-            "meta": OutputPartMeta(
-                partitions=PartitionColumns(
-                    [
-                        PartitionColumn(name="dt", value="2024-01-03"),
-                        PartitionColumn(name="chain", value="op"),
-                    ]
-                ),
-                row_count=1,
+            "partitions": PartitionColumns(
+                cols=[
+                    PartitionColumn(name="dt", value="2024-01-03"),
+                    PartitionColumn(name="chain", value="op"),
+                ]
             ),
             "full_path": "dt=2024-01-03/chain=op",
             "nu_rows": 1,
@@ -132,23 +118,20 @@ def test_breakout_partitions_empty():
     for _ in outputs:
         actual.append(
             dict(
-                meta=_.meta,
-                full_path=_.meta.partitions.path,
+                partitions=_.partitions,
+                full_path=_.partitions.path,
                 nu_rows=len(_.df),
             )
         )
 
-    actual.sort(key=lambda x: x["meta"].partitions.path)
+    actual.sort(key=lambda x: x["full_path"])
     assert actual == [
         {
-            "meta": OutputPartMeta(
-                partitions=PartitionColumns(
-                    [
-                        PartitionColumn(name="chain", value="op"),
-                        PartitionColumn(name="dt", value="2023-10-30"),
-                    ]
-                ),
-                row_count=0,
+            "partitions": PartitionColumns(
+                [
+                    PartitionColumn(name="chain", value="op"),
+                    PartitionColumn(name="dt", value="2023-10-30"),
+                ]
             ),
             "full_path": "chain=op/dt=2023-10-30",
             "nu_rows": 0,
