+++ tests/op_coreutils/partitioned/test_output.py
@@ -1,12 +1,8 @@
 import pyarrow as pa
 
-
-from op_analytics.coreutils.partitioned.output import (
-    ExpectedOutput,
-    PartitionColumns,
-    PartitionColumn,
-)
-from op_analytics.coreutils.partitioned.types import PartitionedRootPath, PartitionedMarkerPath
+from op_analytics.coreutils.partitioned.output import ExpectedOutput
+from op_analytics.coreutils.partitioned.partition import PartitionColumn, PartitionColumns
+from op_analytics.coreutils.partitioned.types import PartitionedMarkerPath, PartitionedRootPath
 
 
 def test_01():
